# pb/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pb/service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/containers": {
      "post": {
        "summary": "Like 'docker run' command",
        "description": "Input/Output is a same protobuf/json object. For input:\n{\n--\"config\":\n----{\n------\"image\": \"nginx\",\n------\"exposed_ports\":\n--------{\n----------\"value\": \"webui\"\n--------}\n----},\n--\"host_config\":\n----{\n------\"port_bindings\":\n--------{\n----------\"value\":\n------------{\n--------------\"80\":\n----------------{\n------------------\"host_port\": \"80\"\n----------------}\n------------}\n--------}\n----},\n--\"network_config\":\n----{\n----},\n--\"container_name\": \"nginx\"\n}\nAnd returning information append this object for output:\n{\n--\"state_code\": 0,  // succeeded, otherwise none zero\n--\"state_message\": \"if failed, provide error information\",\n--\"container_id\": \"regturned from docker engine\"  \n}",
        "operationId": "RunContainer",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbDockerRunData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbDockerRunData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    },
...
    "/v1/provisions": {
      "post": {
        "summary": "Run containers with a user namespace information",
        "description": "Input/Output is a same protobuf/json object. For input:\n{\n  \"name\": \"fighter and target\"\n  \"namespace\": \"default\"\n  \"metadata\":\n    {\n      \"categroy_name\": \"basic-web-security\",\n      \"class_name\": \"http-protocol\"\n      \"field_name\": \"http-method\"\n    },\n  \"provisionings\": [\n    \u003clist of DockerRunData type, see previous\u003e\n  ]\n}",
        "operationId": "ProvisionContainers",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbProvisioningsData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbProvisioningsData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    },
    "/v1/pull": {
      "post": {
        "summary": "Like 'docker pull' command",
        "description": "Input/Output is a same protobuf/json object. For input:\n{\n  \"image\": \"tomcat:8\"\n}",
        "operationId": "PullImage",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbDockerPullData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbDockerPullData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    },
    "/v1/reap-brns": {
      "post": {
        "operationId": "ReapBridgedNetworkLandscape",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbBridgedNetworkingData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbBridgedNetworkingData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    },
    "/v1/reap-instantiation": {
      "post": {
        "summary": "Find containers with instantiating of a user namespace",
        "description": "Input/Output is a same protobuf/json object. For input:\n{\n--\"name\": \"fighter and target\"\n--\"namespace\": \"default\"\n--\"metadata\":\n----{\n------\"categroy_name\": \"default\",\n------\"class_name\": \"default\"\n------\"field_name\": \"default\"\n----}\n}\nFor output, plus a returning object:\n{\n--// Objects copy of input\n--\"instantiation\": [\n----{\n------// List of Moby Container type\n----}\n--]\n}",
        "operationId": "ReapInstantiation",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbInstantiationData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbInstantiationData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    },
    "/v1/reap-registry": {
      "post": {
        "summary": "List registry, include Docker Hub, or private registry (using .docker/config.json)",
        "description": "Input/Output is a same protobuf/json object. For input:\n{\n--\"repositories\": [\n----{\n------name: \"127.0.0.1:5000\"\n----}\n--]\n}\nFor output, plus a returning object:\n{\n--// Objects copy of input\n--\"repositories\": [\n----{\n------name: \"127.0.0.1:5000\",\n------catalogs: [\n--------{\n----------\"name\": \"nginx\",\n----------\"tags\": [\n------------{\n--------------\"name\": \"latest\"\n------------}\n----------]\n--------}\n------]\n----}\n--],\n--\"state_code\": 0,  // Value greater than zero indicates error \n--\"state_message\": \"...\" // Usually error message \n}",
        "operationId": "ReapRegistryForRepositories",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbRegistryRepositoryData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbRegistryRepositoryData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    },
    "/v1/terminations": {
      "post": {
        "summary": "Delete containers with a user namespace information",
        "description": "Input/Output is a same protobuf/json object. For input:\n{\n  \"name\": \"fighter and target\"\n  \"namespace\": \"default\"\n  \"metadata\":\n    {\n      \"categroy_name\": \"basic-web-security\",\n      \"class_name\": \"http-protocol\"\n      \"field_name\": \"http-method\"\n    }\n}\nAnd returning information append this object for output:\n{\n  \"provisionings\": [\n    list of DockerRunData type, see previous\n  ]\n}",
        "operationId": "TerminationContainers",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbProvisioningsData"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbProvisioningsData"
            }
          }
        ],
        "tags": [
          "EchoService"
        ]
      }
    }
  },
  "definitions": {
...
    "mobyConfig": {
      "type": "object",
      "properties": {
        "args_escaped": {
          "type": "boolean",
          "format": "boolean"
        },
        "attach_stderr": {
          "type": "boolean",
          "format": "boolean"
        },
        "attach_stdin": {
          "type": "boolean",
          "format": "boolean"
        },
        "attach_stdout": {
          "type": "boolean",
          "format": "boolean"
        },
        "cmd": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "domainname": {
          "type": "string",
          "format": "string"
        },
        "entrypoint": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "env": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "exposed_ports": {
          "$ref": "#/definitions/mobyPortSet"
        },
        "healthcheck": {
          "$ref": "#/definitions/mobyHealthConfig"
        },
        "hostname": {
          "type": "string",
          "format": "string"
        },
        "image": {
          "type": "string",
          "format": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "string"
          }
        },
        "mac_address": {
          "type": "string",
          "format": "string"
        },
        "network_disabled": {
          "type": "boolean",
          "format": "boolean"
        },
        "on_build": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "open_stdin": {
          "type": "boolean",
          "format": "boolean"
        },
        "shell": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "stdin_once": {
          "type": "boolean",
          "format": "boolean"
        },
        "stop_signal": {
          "type": "string",
          "format": "string"
        },
        "stop_timeout": {
          "type": "integer",
          "format": "int32"
        },
        "tty": {
          "type": "boolean",
          "format": "boolean"
        },
        "user": {
          "type": "string",
          "format": "string"
        },
        "volumes": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "string"
          }
        },
        "working_dir": {
          "type": "string",
          "format": "string"
        }
      },
      "title": "Config contains the configuration data about a container.\nIt should hold only portable information about the container.\nHere, \"portable\" means \"independent from the host we are running on\".\nNon-portable information *should* appear in HostConfig.\nAll fields added to this struct must be marked 'omitempty' to keep getting\npredictable hashes from the old 'v1Compatibility' configuration.\nto see https://github.com/moby/moby/blob/master/api/types/container/config.go"
    },
...
    "mobyHostConfig": {
      "type": "object",
      "properties": {
        "auto_remove": {
          "type": "boolean",
          "format": "boolean"
        },
        "binds": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          },
          "title": "Applicable to all platforms"
        },
        "cap_add": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          },
          "title": "Applicable to UNIX platforms"
        },
        "cap_drop": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "cgroup": {
          "type": "string",
          "format": "string"
        },
        "console_size_height": {
          "type": "integer",
          "format": "int64",
          "title": "Applicable to Windows"
        },
        "console_size_width": {
          "type": "integer",
          "format": "int64"
        },
        "container_id_file": {
          "type": "string",
          "format": "string"
        },
        "dns": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "dns_options": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "dns_search": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "extra_hosts": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "group_add": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "init": {
          "type": "boolean",
          "format": "boolean",
          "title": "Run a custom init inside the container, if null, use the daemon's configured settings"
        },
        "ipc_mode": {
          "type": "string",
          "format": "string"
        },
        "isolation": {
          "type": "string",
          "format": "string"
        },
        "links": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "log_config": {
          "$ref": "#/definitions/mobyLogConfig"
        },
        "mounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/mobyVolumeMount"
          },
          "title": "Mounts specs used by the container"
        },
        "network_mode": {
          "type": "string",
          "format": "string"
        },
        "oom_score_adj": {
          "type": "integer",
          "format": "int32"
        },
        "pid_mode": {
          "type": "string",
          "format": "string"
        },
        "port_bindings": {
          "$ref": "#/definitions/mobyPortMap"
        },
        "privileged": {
          "type": "boolean",
          "format": "boolean"
        },
        "publish_all_ports": {
          "type": "boolean",
          "format": "boolean"
        },
        "readonly_rootfs": {
          "type": "boolean",
          "format": "boolean"
        },
        "resources": {
          "$ref": "#/definitions/mobyResources",
          "title": "Contains container's resources (cgroups, ulimits)"
        },
        "restart_policy": {
          "$ref": "#/definitions/mobyRestartPolicy"
        },
        "runtime": {
          "type": "string",
          "format": "string"
        },
        "security_opt": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        },
        "shm_size": {
          "type": "string",
          "format": "int64"
        },
        "storage_opt": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "string"
          }
        },
        "sysctls": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "string"
          }
        },
        "tmpfs": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "string"
          }
        },
        "userns_mode": {
          "type": "string",
          "format": "string"
        },
        "uts_mode": {
          "type": "string",
          "format": "string"
        },
        "volume_driver": {
          "type": "string",
          "format": "string"
        },
        "volumes_from": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string"
          }
        }
      },
      "title": "HostConfig the non-portable Config structure of a container.\nHere, \"non-portable\" means \"dependent of the host we are running on\".\nPortable information *should* appear in Config.\nto see https://github.com/moby/moby/blob/master/api/types/container/host_config.go"
    },
...
    "pbBridgedNetworkingData": {
      "type": "object",
      "properties": {
        "containers_networking": {
          "$ref": "#/definitions/BridgedNetworkingDataContainersNetworkingInfo"
        },
        "linux_bridges": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BridgedNetworkingDataLinuxBridgeInfo"
          }
        },
        "state_code": {
          "type": "integer",
          "format": "int32"
        },
        "state_message": {
          "type": "string",
          "format": "string"
        },
        "veth_pairs": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "string"
          }
        }
      }
    },
    "pbDockerPullData": {
      "type": "object",
      "properties": {
        "image": {
          "type": "string",
          "format": "string"
        },
        "image_id": {
          "type": "string",
          "format": "string"
        },
        "progress_report": {
          "type": "string",
          "format": "string"
        },
        "state_code": {
          "type": "integer",
          "format": "int32"
        },
        "state_message": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "pbDockerRunData": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/mobyConfig"
        },
        "container_id": {
          "type": "string",
          "format": "string"
        },
        "container_name": {
          "type": "string",
          "format": "string"
        },
        "host_config": {
          "$ref": "#/definitions/mobyHostConfig"
        },
        "network_config": {
          "$ref": "#/definitions/mobyNetworkingConfig"
        },
        "state_code": {
          "type": "integer",
          "format": "int32"
        },
        "state_message": {
          "type": "string",
          "format": "string"
        }
      }
    },
...
    "pbInstantiationData": {
      "type": "object",
      "properties": {
        "instantiation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/mobyContainer"
          }
        },
        "metadata": {
          "$ref": "#/definitions/pbInstrumentMetadata"
        },
        "name": {
          "type": "string",
          "format": "string"
        },
        "namespace": {
          "type": "string",
          "format": "string"
        },
        "state_code": {
          "type": "integer",
          "format": "int32"
        },
        "state_message": {
          "type": "string",
          "format": "string"
        }
      }
    },
...
    "pbProvisioningsData": {
      "type": "object",
      "properties": {
        "metadata": {
          "$ref": "#/definitions/ProvisioningsDataMetadata"
        },
        "name": {
          "type": "string",
          "format": "string"
        },
        "namespace": {
          "type": "string",
          "format": "string"
        },
        "provisionings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbDockerRunData"
          }
        }
      }
    },
    "pbRegistryRepositoryData": {
      "type": "object",
      "properties": {
        "registries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RegistryRepositoryDataRegistry"
          }
        },
        "state_code": {
          "type": "integer",
          "format": "int32"
        },
        "state_message": {
          "type": "string",
          "format": "string"
        }
      }
    }
  }
}